requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.2.0
audioread==2.1.9
backcall==0.2.0
bleach==3.2.1
bob.bio.base==4.1.1
bob.bio.gmm==3.2.6
bob.blitz==2.0.22
bob.core==2.2.6
bob.db.base==3.0.5
bob.extension==5.0.1
bob.io.audio==2.0.13
bob.io.base==3.0.9
bob.io.image==2.4.6
bob.io.matlab==2.0.17
bob.kaldi==1.3.0
bob.learn.activation==2.0.16
bob.learn.em==2.1.6
bob.learn.linear==2.1.8
bob.math==3.1.7
bob.measure==4.2.2
bob.sp==2.0.16
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
confuse==1.3.0
cycler==0.10.0
dataclasses==0.7
decorator==4.4.2
defusedxml==0.6.0
deprecation==2.1.0
docopt==0.6.2
entrypoints==0.3
eyeD3==0.8.12
filetype==1.0.7
hmmlearn==0.2.2
idna==2.10
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1604014557000/work
librosa==0.8.0
llvmlite==0.34.0
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.0.4
mkl-service==2.3.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
notebook==6.1.4
numba==0.51.2
numpy==1.18.1
olefile==0.46
packaging==20.4
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///tmp/build/80754af9/pillow_1603822238230/work
plotly==4.12.0
pooch==1.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
-e git+https://github.com/tyiannak/pyAudioAnalysis.git@7239d78c04a1c72b2ef85af0e21e0b451c3c8dd7#egg=pyAudioAnalysis
pycparser==2.20
pydub==0.23.1
Pygments==2.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-magic==0.4.18
python-speech-features==0.6
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.24.0
resampy==0.2.2
retrying==1.3.3
scikit-learn==0.21.3
scipy==1.4.1
Send2Trash==1.5.0
simplejson==3.16.0
six==1.15.0
SoundFile==0.10.3.post1
SQLAlchemy @ file:///tmp/build/80754af9/sqlalchemy_1603398008024/work
tabulate==0.8.7
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.44.1
traitlets==4.3.3
urllib3==1.25.11
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.4.0
